trie: partial checkpointing fix
